Replaced the boolean shorten pauses setting with an expanded setting that allows to shorten no pauses, pauses at utterance boundaries, or all punctuated pauses. (#101)
To implement this, the shortPause driver setting has been replaced with a choice setting called pauseMode, using logic based on that used for sample rate selection.
This introduces a third pause mode, herein called utterance boundary, that will only reduce the pause at the end of a speech sequence. This has the value 1.
The value 2 corresponds to the previous shorten pauses setting set to true, causing the utterance boundary to be shortened as well as the pause-related regular expressions to be applied.
This does not presently load the previous shorten pauses setting and adapt to this new scheme, and the string for the utterance mode could likely do with some revision.